+++ src/powermouse/domain/controllers/camera.py
@@ -5,9 +5,6 @@ class CameraController:
     def __init__(self, camera: Camera):
         self.camera = camera
 
-    def list_cameras(self) -> list[Camera]:
-        raise NotImplementedError
-
     def update_frame(self):
         raise NotImplementedError