+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.0.2
+
+- Updated fonts for readability.
+- Camera names are added to the onboarding and camera widget dropdowns.
+
 ## v0.0.1
 
 - Alpha testing release.